engine, dsp: the sample clock, the profile cache, and noise without transcendentals
Three fixes for the eyes-on's remaining reports, planned in
docs/waveform-performance-plan.md and benchmark-gated:

- The observer walks a sample clock. A millisecond is 62.5 samples at
  62.5 kHz, and truncating the half in StartSample tore the stream's
  phase at every window seam - forty broadband clicks a second, drawn
  as a burst filling the whole span. rxTransmission takes a fractional
  anchor now and carries the sub-sample remainder on DelaySamples,
  where rf.Observe turns it into the phase it is; ObserveSignalAt
  addresses the stream by absolute sample index.

- A geometry-keyed profile cache, and probably the field stall itself:
  MeshCore firmware effectively issues a radio report around every
  transmission via its FEM/AGC fields, each report invalidated the
  pair's loss, and recomputing a loss re-walked up to 257 DEM lookups
  synchronously inside the judge path. With several transmitters on
  real terrain that is 'stutters until it pretty much stops'. The
  profile now outlives every radio report - a test holds a report to
  zero DEM lookups - and dies only when a node actually moves.

- Box-Muller replaced by Acklam's inverse normal CDF (18.75 ns/pair,
  tails held by test - the FEC threshold lives in the tails), and the
  FFT got per-size twiddle and bit-reversal plans. Seeded noise
  realisations moved once; every outcome-pinned test passed unchanged.
  Engine bench: 8 concurrent talkers 199 -> 152 ms per 5 s simulated,
  3 talkers 113 -> 55 ms, with GPU synthesis planned next for the
  remainder.

+++ internal/engine/waveform.go
@@ -237,7 +237,7 @@ func (e *Engine) judgeWaveform(t transmission, c wfCandidate, concurrent []trans
 		if !e.phyOf(nodes[other.from].Spec).sameChannel(txPHY) {
 			continue
 		}
-		txs = append(txs, e.rxTransmissions(other, c.i, t.startMs, nodes, cache)...)
+		txs = append(txs, e.rxTransmissions(other, c.i, float64(t.startMs), nodes, cache)...)
 	}
 
 	noiseLinear := math.Pow(10, e.applyImplementationLoss(c.noiseDBm)/10)
@@ -388,7 +388,7 @@ func (e *Engine) waveformBusy(now uint32, nodes []*Node, air []transmission) []b
 			if !e.phyOf(nodes[t.from].Spec).sameChannel(rxPHY) {
 				continue
 			}
-			if tx, ok := e.rxTransmission(t, i, now, nodes, cache); ok {
+			if tx, ok := e.rxTransmission(t, i, float64(now), nodes, cache); ok {
 				txs = append(txs, tx)
 			}
 		}
